fastldpc: il contatore dei candidati a 64 bit, allineato al laboratorio
Su Windows (LLP64) long e' a 32 bit, e con la ricerca larga i candidati
sottoposti alla CRC superano i 2 miliardi: il contatore diventava negativo. In
esercizio e' dietro -DOSD_COUNT, che non e' definito, quindi il codice e'
compilato via e non cambia niente di funzionante -- e' una trappola tolta per
chi domani lo accenda per misurare.

Gli header di esercizio e quelli del laboratorio tornano cosi' identici, tranne
una differenza VOLUTA: l'esercizio ha la costante singola alpha_w, il
laboratorio l'API dei pesi per iterazione piu' quelli per arco del pacchetto
neurale. Il laboratorio e' un soprainsieme, ed e' la relazione giusta; verificato
che a peso uniforme diano gli stessi numeri (11964 e 12371 su v0.5).

+++ Detector/fastldpc/ft2_decoder.hpp
@@ -95,7 +95,7 @@ public:
     // da cui dipendono i nominativi fantasma: la CRC ne ammette uno ogni 16384.
     // E' strutturale e deterministico, quindi dice quello che il conteggio dei
     // fantasmi su una piscina di rumore gaussiano non riesce a dire.
-    long crc_tests() const { return osd_.n_crc; }
+    long long crc_tests() const { return osd_.n_crc; }
     void reset_crc_tests() { osd_.n_crc = 0; }
 
     // apmask (opzionale): [n][174], 1 sui bit gia' noti per ipotesi a priori.
